requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
reportlab>=4.0.0
streamlit>=1.30.0
flask
flask-cors
requests
